features added: add new items to existing order, orders list and detail view for users, refactor

+++ webshop/src/main/java/com/e_commerce/webshop/dto/ordersbyuser/OrdersByUserProductQuantityDTO.java
@@ -4,11 +4,13 @@ import com.e_commerce.webshop.model.ProductQuantity;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class OrdersByUserProductQuantityDTO {
     private Long id;
-    private int sum;
+    private BigDecimal sum;
     private int quantity;
     private OrdersByUserProductDTO product;
     public OrdersByUserProductQuantityDTO(ProductQuantity productQuantity) {
